testing

+++ client.py
@@ -16,7 +16,6 @@ except:
     pass
 
 client.bind(clientsock)
-client.setblocking(0)
 client.connect(hubsocket)
 clientfd=client.fileno()
 
@@ -29,18 +28,25 @@ while 1:
                 client_packet=os.read(0,2)
                 packet_length=(ord(client_packet[:1:])*256)+ord(client_packet[1:2:])
                 client_packet=os.read(0,packet_length)
-                if not client_packet:
-                    os.remove(clientsock)
-                    break
-                packet_length=len(client_packet)
-                try:
-                    write_length=client.sendto(client_packet,hubsocket)
-                    if write_length!=packet_length:
-                        os.write(2,'error: write_length == '+str(write_length)+', packet_length == '+str(packet_length)+'\n')
-                except:
-                    os.write(2,'error: cannot write to '+hubsocket+'\n')
+                if client_packet!=packet_length:
+                    while client_packet!=packet_length:
+                        client_packet+=os.read(0,packet_length-client_packet)
+                        if not client_packet:
+                            os.remove(clientsock)
+                            break
             except:
                 os.write(2,'error: udpmsg4 protocol error\n')
+
+            if not client_packet:
+                os.remove(clientsock)
+                break
+            packet_length=len(client_packet)
+            try:
+                write_length=client.sendto(client_packet,hubsocket)
+                if write_length!=packet_length:
+                    os.write(2,'error: write_length == '+str(write_length)+', packet_length == '+str(packet_length)+'\n')
+            except:
+                os.write(2,'error: cannot write to '+hubsocket+'\n')
 
         if clientfd in read_this:
             hub_packet=client.recv(65536)
